Add ArkadCheckbox component, show days at company profile, update design on EditUserProfile
